REFACTOR(All): Move database folder
