+++ 0_Game_System/tsconfig.json
@@ -27,7 +27,7 @@
       "@components*": ["./src/components/*"],
       "@pages*": ["./src/components/pages/*"],
       "@interfaces*": ["./src/components/interfaces/*"],
-      "@database*": ["./src/components/database/*"],
+      "@database*": ["./src/components/services/database/*"],
       "@UI*": ["./src/components/UI/*"],
       "@Icons*": ["./src/components/UI/Icons/*"],
       "@utils*": ["./src/components/utils/*"],
